+++ src/games/candy-crush/const/const.ts
@@ -1,0 +1,29 @@
+/**
+ * @author       Digitsensitive <digit.sensitivee@gmail.com>
+ * @copyright    2020 digitsensitive
+ * @description  Candy crush: Const
+ * @license      Digitsensitive
+ */
+
+export let CONST = {
+  score: 0,
+  highscore: 0,
+  gridWidth: 8,
+  gridHeight: 8,
+  tileWidth: 64,
+  tileHeight: 72,
+  candyTypes: [
+    "cookie1",
+    "cookie2",
+    "croissant",
+    "cupcake",
+    "donut",
+    "eclair",
+    "macaroon",
+    "pie",
+    "poptart1",
+    "poptart2",
+    "starcookie1",
+    "starcookie2"
+  ]
+};