+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Add new `experimental` project `point-in-polygon` *[Author: digitsensitive (2020-03-08)]*
 * Add new `experimental` project `fractal-tree` *[Author: digitsensitive (2020-03-08)]*
 * Add new `experimental` project `lissajous-curve` *[Author: digitsensitive (2020-03-21)]*
+* Add new game `candy-crush` *[Author: digitsensitive (2020-03-26)]*
 
 ### Updates